Fixed problem in the init script
Updated the Cassandra init script so that the `cassandra` service starts properly and shuts down cleanly.

The current script repeats most of what's in `/usr/local/cassandra/bin/cassandra`, but not enough, there's a lot of the environment Cassandra needs to start that is missing. This patch changes it so that it simply runs `cassandra -p $PIDFILE` and daemonizes it with `start-stop-daemon` (by way of `start_daemon`). The stop and status tasks had to change too.