Fixes broken version control via chef node attribute setting. Updates default setting to Cassandra version 1.2.5.
When trying to set the version on the cassandra tarball via node[:cassandra][:version] this wasn't recognized due to the local variable cassandra_version in the default attributes of the cassandra cookbook that overrides any such outer setting. However this is fixed now by removing that variable and setting the tarball configuration after version setting.