Add slash to file between path components
If slash is not present in the attribute then the concatenation produces
a path where the `cassandra` user has no privileges to write.

When the init file runs it touches the pid file, then cassandra fails to
start, then all subsequent times the init file is invoked, the pid file
exists and no pid is inside so `is_running` fails always.

I prefer to put the slash in this template so the attribute can be
specified in any way. Two slashes work, none not.