Move installation directory creation into the debian section.
Shouldn't break anything on the debian platform if it is required but it
can cause problems on rhel - especially if you've foolishly pointed
conf_dir at /etc/cassandra/conf which is where the package wants to put
a symlink.

Not sure why these are needed for packaged installs, but if they are, I
suspect at least some of them should be done after the package is
installed.

I've moved conf_dir down to ensure that it sets permissions
appropriately whether it is used by the package or not.